credit_card_chatbot_site/backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.4
python-dotenv==1.0.0
